telegram_bot/requirements.txt
python-telegram-bot
git+https://github.com/okeer/numdl.git@master#numdl
